suivi complete

+++ app/src/main/java/com/example/gtm/data/entities/response/Response.kt
@@ -6,7 +6,7 @@ data class Response(
     val id: Int,
     val question: QuestionX,
     val questionId: Int,
-    val rate: Int,
+    val rate: Float,
     val responsePictures: List<ResponsePicture>,
     val surveyResponseId: Int,
     val updatedAt: String
